requirements.txt
streamlit
yfinance
python-bcb
pandas
numpy
plotly